+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+-   [ENH] Added `rle_id` function for run-length encoding IDs - Issue #1435 @emmanuel-ferdman
+
 ## [v0.32.3] - 2025-12-11
 
 -   [ENH] Undeprecate `select_columns` and `select_rows` - Issue #1514 @samukweku